# Ignore files in .gitignore in tidy