@jyn514 2023-01-04
I originally tried making all these checks parallel before realizing that you'd already done that at the top level 🤦 just not at the file level which is the thing I did. If you want the code it's here: https://github.com/rust-lang/rust/compare/master...jyn514:rust:parallel-tidy?expand=1)

but not sure how useful it is given that `walk` is within the checks and not at the top-level, seems pretty hard to be able to intersperse the checks with each other.