@the8472 2023-01-04
Switching to file level parallelism may be better on many-core systems because doing it at the top level is fairly coarse and the longest-running tasks (`style.rs` was the slowest last time I checked) end up dominating wall-time.
I did the top-level parallelization because it was good enough back then, but tidy has grown since.

I'll do a comparison of this branch vs. the file-parallel one on my machine.