@jyn514 2023-01-04
oh sorry I forgot to mention, it saved nearly no time at all :/ first image is with this PR and second image is with file-level parallelism.
![image](https://user-images.githubusercontent.com/23638587/210491805-e77253d9-fa21-4c8d-902a-c453ae8f67a3.png)

![image](https://user-images.githubusercontent.com/23638587/210491815-93b38dff-fdd6-4801-aa96-8ce49c913029.png)