@jyn514 2023-01-04
Oh, interesting - I think all the cargo calls are in `src/tools/tidy/src/deps.rs`, which is why they're serialized - I think it shouldn't be too hard to `spawn` them all in parallel and then wait on each :)

I don't want to block this PR on that change though, I think it's already a good improvement even without the file-level parallelism.