@jyn514 2023-03-05
It's used in `mir_opt_tests`: https://github.com/rust-lang/rust/blob/40ba0e84d53f605ccf01836e9c2d27892728ae81/src/tools/tidy/src/mir_opt_tests.rs#L9