@jyn514 2023-03-05
I don't think so, the only files there are `.rs`, `.mir`, `.diff`, and `.html` files, and all of them are looked at (intentionally).

@bors r=the8472 rollup